pyproject.toml
[project]
name = "rfdetrtraining"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-cloud-aiplatform>=1.132.0",
    "google-cloud-storage>=3.7.0",
    "hypertune>=1.2.1",
    "python-dotenv>=1.2.1",
    "setuptools>=80.9.0",
]
